extracted searchfunction to its own component ? extracted from combinedsearchfilterbar.js
