+++ src/components/SearchBar.js
@@ -1,37 +1,90 @@
-// src/components/SearchBar.js
-import React, { useState } from 'react';
+import React from 'react';
+import { styled, alpha } from '@mui/material/styles';
+import InputBase from '@mui/material/InputBase';
+import SearchIcon from '@mui/icons-material/Search';
+import IconButton from '@mui/material/IconButton';
+import MenuIcon from '@mui/icons-material/Menu';
+import { Box } from '@mui/material';
 
-const SearchBar = ({ onSearch }) => {
-    const [searchTerm, setSearchTerm] = useState('');
+const Search = styled('div')(({ theme }) => ({
+  position: 'relative',
+  borderRadius: theme.shape.borderRadius,
+  backgroundColor: alpha(theme.palette.common.white, 0.15),
+  '&:hover': {
+    backgroundColor: alpha(theme.palette.common.white, 0.25),
+  },
+  marginLeft: 0,
+  width: '100%',
+  [theme.breakpoints.up('sm')]: {
+    marginLeft: theme.spacing(1),
+    width: 'auto',
+  },
+}));
 
-    const handleSearchChange = (event) => {
-        setSearchTerm(event.target.value);
-    };
+const SearchIconWrapper = styled('div')(({ theme }) => ({
+  padding: theme.spacing(0, 2),
+  height: '100%',
+  position: 'absolute',
+  pointerEvents: 'none',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+}));
 
-    const handleSearchSubmit = (event) => {
-        event.preventDefault();
-        onSearch(searchTerm); 
-    };
+const StyledInputBase = styled(InputBase)(({ theme }) => ({
+  color: 'inherit',
+  width: '100%',
+  '& .MuiInputBase-input': {
+    padding: theme.spacing(1, 1, 1, 0),
+    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+    transition: theme.transitions.create('width'),
+    [theme.breakpoints.up('sm')]: {
+      width: '12ch',
+      '&:focus': {
+        width: '20ch',
+      },
+    },
+  },
+}));
 
-    return (
-        <form onSubmit={handleSearchSubmit}>
-            <input
-                type="text"
-                placeholder="Search Pokémon..."
-                value={searchTerm}
-                onChange={handleSearchChange}
-                style={{
-                    padding: '10px',
-                    fontSize: '16px',
-                    width: '100%',
-                    boxSizing: 'border-box',
-                    borderRadius: '5px',
-                    border: '1px solid #ccc',
-                }}
-            />
-            <button type="submit" style={{ display: 'none' }}>Search</button>
-        </form>
-    );
+const SearchBar = ({ searchTerm, setSearchTerm, onSearch, onDrawerOpen }) => {
+  const handleSearchChange = (event) => {
+    setSearchTerm(event.target.value);
+  };
+
+  const handleSearchSubmit = (event) => {
+    event.preventDefault();
+    onSearch(searchTerm);
+  };
+
+  return (
+    <Box sx={{ display: 'flex', alignItems: 'center', width: '100%' }}>
+      <IconButton
+        size="large"
+        edge="start"
+        color="inherit"
+        aria-label="open drawer"
+        onClick={onDrawerOpen}
+        sx={{ mr: 2 }}
+      >
+        <MenuIcon />
+      </IconButton>
+      
+      <form onSubmit={handleSearchSubmit} style={{ flex: 1, display: 'flex', alignItems: 'center' }}>
+        <Search sx={{ borderStyle: 'none', width: '100%' }}>
+          <SearchIconWrapper>
+            <SearchIcon />
+          </SearchIconWrapper>
+          <StyledInputBase
+            placeholder="Search…"
+            value={searchTerm}
+            onChange={handleSearchChange}
+            inputProps={{ 'aria-label': 'search' }}
+          />
+        </Search>
+      </form>
+    </Box>
+  );
 };
 
 export default SearchBar;
